% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCelltypeList.R
\name{getCelltypeList}
\alias{getCelltypeList}
\title{getCelltypeList}
\usage{
getCelltypeList(protocol, species)
}
\arguments{
\item{protocol}{: 1-DNase-seq(ENCODE),2-ATAC-seq(ENCODE),3-ATAC-seq(ATACdb)}

\item{species}{: genome version of human of mouse, 11-hg19,12-hg38,21-mm9,22-mm10}
}
\description{
Get the corresponding cell type list
}
\examples{
getCelltypeList(1,1)
}
