% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExampleInputFile.R
\name{getExampleInputFile}
\alias{getExampleInputFile}
\title{getExampleInputFile}
\usage{
getExampleInputFile(save_path)
}
\value{
save_path : Path to save download file.
}
\description{
Download example input file to the local path
}
\examples{
getExampleInputFile('.')
}
