% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExampleTaskID.R
\name{getExampleTaskID}
\alias{getExampleTaskID}
\title{getExampleTaskID}
\usage{
getExampleTaskID()
}
\value{
task_id
}
\description{
Get an example task id
}
\examples{
task_id <- getExampleTaskID()
}
