% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInputFile.R
\name{getInputFile}
\alias{getInputFile}
\title{getInputFile}
\usage{
getInputFile(save_path, task_id)
}
\arguments{
\item{save_path}{: Path to save download file.}

\item{task_id}{: The 16-bit identity of the submitted task.}
}
\description{
Download input file to the local path
}
\examples{
getInputFile('.',2020121013091517)
}
