% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProgress.R
\name{getProgress}
\alias{getProgress}
\title{getProgress}
\usage{
getProgress(task_id)
}
\arguments{
\item{task_id}{: The 16-bit identity of the submitted task.}
}
\description{
You can view the annotation progress
}
\examples{
getProgress(2020121013091517)
}
