% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSystemList.R
\name{getSystemList}
\alias{getSystemList}
\title{getSystemList}
\usage{
getSystemList(protocol, species)
}
\arguments{
\item{protocol}{: 1-DNase-seq(ENCODE),2-ATAC-seq(ENCODE),3-ATAC-seq(ATACdb)}

\item{species}{: genome version of human of mouse, 1-hg19,2-hg38,3-mm9,4-mm10}
}
\description{
Get the corresponding tissue list
}
\examples{
getSystemList(1,11)
}
