% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Openness2Seurat.R
\name{Openness2Seurat}
\alias{Openness2Seurat}
\title{Openness2Seurat}
\usage{
Openness2Seurat(file_path)
}
\arguments{
\item{file_path}{: Path of download file.}
}
\description{
Download the annotation result
}
