% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnnotate.R
\name{runAnnotate}
\alias{runAnnotate}
\title{runAnnotate}
\usage{
runAnnotate(file_path, species, protocol, perbase)
}
\arguments{
\item{file_path}{: The path of the '.bed' or '.bed.gz' file to be uploaded.}

\item{species}{: genome version of human of mouse, 11-hg19,12-hg38,21-mm9,22-mm10}

\item{protocol}{: 1-DNase-seq(ENCODE),2-ATAC-seq(ENCODE),3-ATAC-seq(ATACdb)}

\item{perbase}{: 0 : Region based,1 : Per-base based.}
}
\value{
taskID
}
\description{
Upload .bed or .bed.gz file to the server
}
