% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchCelltype.R
\name{searchCelltype}
\alias{searchCelltype}
\title{searchCelltype}
\usage{
searchCelltype(protocol, species, keyword)
}
\arguments{
\item{protocol}{: 1-DNase-seq(ENCODE),2-ATAC-seq(ENCODE),3-ATAC-seq(ATACdb)}

\item{species}{: genome version of human of mouse, 11-hg19,12-hg38,21-mm9,22-mm10}

\item{keyword}{: key word you want to search}
}
\description{
Search for the corresponding cell type list
}
\examples{
searchCelltype(1,11,'K562')
}
