% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchTissue.R
\name{searchTissue}
\alias{searchTissue}
\title{searchTissue}
\usage{
searchTissue(protocol, species, keyword)
}
\arguments{
\item{protocol}{: 1-DNase-seq(ENCODE),2-ATAC-seq(ENCODE),3-ATAC-seq(ATACdb)}

\item{species}{: genome version of human of mouse, 1-hg19,2-hg38,3-mm9,4-mm10}

\item{keyword}{: key word you want to search}
}
\description{
Search for the corresponding cell type list
}
\examples{
searchTissue(1,1,'blood')
}
