% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewParams.R
\name{viewParams}
\alias{viewParams}
\title{viewParams}
\usage{
viewParams(task_id)
}
\arguments{
\item{task_id}{: The 16-bit identity of the submitted task.}
}
\description{
You can view the parameters you set before
}
\examples{
viewParams(2020121013091517)
}
